feat(dtfs2-6892): move dtfs central services to tfm api
