+++ dtfs-central-api/src/repositories/tfm-users-repo/tfm-users.repo.ts
@@ -1,6 +1,5 @@
 import { Collection, WithoutId, ObjectId } from 'mongodb';
-import { AuditDetails, MONGO_DB_COLLECTIONS, TfmUser, UserUpsertRequest } from '@ukef/dtfs2-common';
-import { generateAuditDatabaseRecordFromAuditDetails } from '@ukef/dtfs2-common/change-stream';
+import { MONGO_DB_COLLECTIONS, TfmUser } from '@ukef/dtfs2-common';
 import { mongoDbClient } from '../../drivers/db-client';
 
 export class TfmUsersRepo {
@@ -20,24 +19,5 @@ export class TfmUsersRepo {
   public static async findOneUserById(id: string | ObjectId): Promise<TfmUser | null> {
     const collection = await this.getCollection();
     return collection.findOne({ _id: new ObjectId(id) });
-  }
-
-  /**
-   * Upserts a user
-   * @param upsertUserRequest
-   * @returns upserted user
-   */
-  public static async upsertUser({ userUpdate, auditDetails }: { userUpdate: UserUpsertRequest; auditDetails: AuditDetails }): Promise<void> {
-    const collection = await this.getCollection();
-
-    const userUpsert = {
-      ...userUpdate,
-      auditRecord: generateAuditDatabaseRecordFromAuditDetails(auditDetails),
-    };
-
-    const query = { name: 'To Update' }; // TODO: DTFS2-6892: This should be updated as part of this ticket
-    const update = { $set: userUpsert };
-    const options = { upsert: true };
-    await collection.updateOne(query, update, options); // TODO: DTFS2-6892: Test this fails if there are multiple users
   }
 }
