+++ trade-finance-manager-api/src/v1/services/user.service.ts
@@ -1,8 +1,14 @@
-import { EntraIdUser, UserUpsertRequest } from '@ukef/dtfs2-common';
+import { AuditDetails, EntraIdUser, UserUpsertRequest } from '@ukef/dtfs2-common';
 import { ENTRA_ID_USER_TO_TFM_UPSERT_REQUEST_SCHEMA } from '@ukef/dtfs2-common/schemas';
+import { UserRepo } from '../repo/user.repo';
 
 export class UserService {
-  public static transformEntraUserToTfmUserUpsert(entraUser: EntraIdUser): UserUpsertRequest {
+  private static transformEntraUserToTfmUserUpsert(entraUser: EntraIdUser): UserUpsertRequest {
     return ENTRA_ID_USER_TO_TFM_UPSERT_REQUEST_SCHEMA.parse(entraUser);
+  }
+
+  public static async upsertUserFromEntraIdUser({ entraUser, auditDetails }: { entraUser: EntraIdUser; auditDetails: AuditDetails }): Promise<void> {
+    const userUpsertRequest = this.transformEntraUserToTfmUserUpsert(entraUser);
+    await UserRepo.upsertUser({ userUpdate: userUpsertRequest, auditDetails });
   }
 }
